feat: fine tune gradient for dijkstra
Centerlines are much better using the PDRF alone but occasionally
using only the PDRF without the DAF added results in very low value
regions where the gradient collapses into floating point errors.

The intuition here is that using the naked DAF overpowers the PDRF
and distorts the centerline, but only a small gradient is necessary
to correct collapsed regions. Therefore, we normalize the DAF so that
it provides a continuous small gradient that won't overpower the PDRF.