feat: made remove_ticks slower but more accurate
remove_ticks was taking off the ends of some neurites. This was
due to assessing the ends of the neurites from a minor branch point,
e.g. of a tick, which put it under the threshold. Now we remove
the smallest tick, then recompute everything to ensure that the
topology remains intact.