fix: close the stargates
Large loops get fused at a centroid which may not be located
inside the loop proper. This creates a radial structure with
many edges from the central vertex to the loop that I call a
"stargate".

If the distance from a selected central
vertex to the outer ring is greater than the EDT radius of the
central vertex, cut the ring using a single edge.

Edge selection could be done in a more intelligent way, but this
is better than the stargates.