fix: reduce influence of integer truncation on averaged downsamples
Iterative downsample_with_averaging has the advantage that each level
of downsampling is a fourth the size of the original mip level. However,
as most image datasets are integer types, we were previously losing up
to a unit of luminance on each mip level due to integer truncation.

While this is difficult to avoid for superdownsamples, we can avoid it
while maintaining this desirable logarithmic time complexity by averaging
floats and writing integers at each mip level, limiting the total loss
of pixel luminance to at most 1 across all mip levels in a given task.