fix(Mesher): save bit of precision in z
In 32-bit, single bits of precision matter as we only have 10 bits
for z, and 11 for x and y. In ZI_MC_QUICK_INTERP the vertex position
is averaged as (v1 + v2) >> 1, however v1+v2 can overflow, so we do
(v1 >> 1) + (v2 >> 1) instead.

Additionally, I am now super suspicious of ints that are not uints
as they lose a bit of valuable precision so that's scrapped.