feat: add support for shard fragment mesh generation
This also pushes upload of individual labels to the end of the
procedure. This has the potential to increase memory consumption
as mesh binaries will need to be stored, but it will reduce the
tendancy for certain tasks that blow up memory to restart in a
loop and write mesh fragments over and over again which
can be potentially expensive.

If necessary, we can add a flag for online generation.