Reduce read requests when creating sharded skeletons and meshes (#156)
* Cache spatial info files

Because igneous download all spatial info file for every tasks, with
fixed number of segments per shard, the downloading request grows
quadratically with volume. This can blow up the cost for large dataset
if we are charged for read operations. Cache it so it only grows
linearly with volume.

* Add a frag_path parameter to save frag files somewhere else

To create shard files, we at least have to read each segment once. With
billions of segments this can be significant. Add a frag_path parameter
so we can save the fragment files locally or somewhere the read requests
are cheap.

* Add a cache option to control CloudVolume cache behavior

Cache is disabled by default